fix(cli): skip trashed assets in the repair, show the trash row, tighten --limit
Review round 1 of #450.

- The trash fact is exposed once (`Workspace::is_trashed`) and used twice:
  `capsule repair capture-time` skips an asset in trash before reading its
  original and reports the count as its own category, because appending
  an irreversible signed record to an asset the user discarded is not a
  repair; `capsule show` regains its "In trash" row through the same
  accessor, so the catalog contexts that described it describe something
  that exists again.
- `--limit` is a positive count (`value_parser!(u64).range(1..)`), and
  `--limit` without `--apply` is refused with a line naming both flags
  before the library is opened, instead of silently doing nothing.
- A derive/matches mismatch in `run()` is formatted through the localized
  command before it exits, the way clap reports its own errors.
- `ShowError`'s `Display` strings are developer-facing (selector and count
  only); the user always sees `describe_error`, which goes through the
  catalog.
- The help invariant test now also asserts the converse: every `cli.help.*`
  key in the canonical catalog is one the command walk produces, so a key
  orphaned by a rename fails the build.
- Tests: a trashed asset is neither corrected nor counted as affected and
  is picked up again after a restore; `apply` writes exactly
  `Affected::recovered` (tripwire) and the detect → apply path lands the
  fixture's EXIF instant itself; a swept asset shows `In trash: yes` in
  process and through the binary; `--limit 0` and `--limit` without
  `--apply` are refused through the binary.

Catalogs and cli-surface.json regenerated (`mise run i18n`, `mise run
cli-surface`); the `--limit` help text's catalog entry follows its doc comment.

+++ capsule-cli/src/cli/commands.rs
@@ -239,9 +239,9 @@ pub(crate) enum RepairCommands {
         #[arg(long)]
         apply: bool,
         /// Correct at most this many affected assets in one --apply run (the report still
-        /// covers the whole library)
-        #[arg(long, value_name = "COUNT")]
-        limit: Option<usize>,
+        /// covers the whole library); only meaningful with --apply
+        #[arg(long, value_name = "COUNT", value_parser = clap::value_parser!(u64).range(1..))]
+        limit: Option<u64>,
     },
 }
